requirements.txt
PyYAML == 6.0
GitPython == 3.1.27